+++ src/server/network/NetworkServer.hpp
@@ -406,7 +406,7 @@ class NetworkServer {
      * @param callback Function receiving (userId, inputMask)
      */
     void onClientInput(
-        std::function<void(std::uint32_t userId, std::uint8_t input)> callback);
+        std::function<void(std::uint32_t userId, std::uint16_t input)> callback);
 
     /**
      * @brief Register callback for get users request
@@ -644,7 +644,7 @@ class NetworkServer {
     std::function<void(std::uint32_t)> onClientConnectedCallback_;
     std::function<void(std::uint32_t, network::DisconnectReason)>
         onClientDisconnectedCallback_;
-    std::function<void(std::uint32_t, std::uint8_t)> onClientInputCallback_;
+    std::function<void(std::uint32_t, std::uint16_t)> onClientInputCallback_;
     std::function<void(std::uint32_t)> onGetUsersRequestCallback_;
     std::function<void(std::uint32_t, bool)> onClientReadyCallback_;
     std::function<void(std::uint32_t, const std::string&)>
